Reorganize into libraries
This should make it easier to reuse the code later, for example for ROS.
