+++ aruco/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.16...3.31)
-project(ArucoSimulation VERSION 1.0
-                        DESCRIPTION "Try to find boxes in a simulated environment"
-                        LANGUAGES CXX)
+project(BananasArucoDetector VERSION 1.0
+                             DESCRIPTION "Try to find boxes using ArUco markers"
+                             LANGUAGES CXX)
 
 # OpenCV depends on the deprecated FindCUDA for some reason
 cmake_policy(SET CMP0146 OLD)
@@ -28,25 +28,5 @@ find_package(OGRE REQUIRED COMPONENTS Bites CONFIG)
 # In case someone out there is still using a system with libm
 find_library(MATH_LIBRARY m)
 
-add_executable(generate_board generate_board.cpp)
-target_compile_features(generate_board PUBLIC cxx_std_17)
-set_target_properties(generate_board PROPERTIES CXX_EXTENSIONS OFF)
-target_compile_options(generate_board PRIVATE -Wall -Wextra -Wpedantic)
-target_link_libraries(generate_board PRIVATE ${OpenCV_LIBS})
-
-add_executable(positioner
-    affine_rotation.h
-    affine_rotation.cpp
-    positioner.cpp
-    visualization/visualizer.h
-    visualization/visualizer.cpp
-    world.h
-    world.cpp
-)
-target_compile_features(positioner PUBLIC cxx_std_17)
-set_target_properties(positioner PROPERTIES CXX_EXTENSIONS OFF)
-target_compile_options(positioner PRIVATE -Wall -Wextra -Wpedantic)
-target_link_libraries(positioner PRIVATE ${OpenCV_LIBS} Microsoft.GSL::GSL OgreBites Eigen3::Eigen)
-if(MATH_LIBRARY)
-    target_link_libraries(positioner PRIVATE ${MATH_LIBRARY})
-endif()
+add_subdirectory(src)
+add_subdirectory(apps)
