+++ aruco/apps/generate_board.cpp
