+++ aruco/include/bananas_aruco/affine_rotation.h
