+++ aruco/src/affine_rotation.cpp
@@ -1,4 +1,4 @@
-#include "affine_rotation.h"
+#include <bananas_aruco/affine_rotation.h>
 
 #include <utility>
